unificacao dos testes de filtrar despesas em um so arquivo
